TimingFunctions
Calculates the average time certain function (adding to the front of a list, adding to a dictionary, deleting from a dictionary, etc.) take to execute. A list can be filled with 1 item to 10000000 items. Once the list size is predetermined, the function is performed the amount of times it takes for the empty list to be filled with the predetermined number. We repeat this process 10 times.